refactor rpc dispatching
